tendenci/project_template/requirements/prod.txt
-r common.txt
# The following packages are needed for Heroku.
# You may not need them locally.

# Memcache
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4

# For Sentry
raven==2.0.2
